Move axios setup out of app.js
